Add code to check if agency name in current-federal matches CyHy ID
There is at least one case (AMTRAK) where an agency is not in the
agency mapping file (agencies.csv) because its name as listed in
current-federal.csv is identical to its CyHy ID.  (See AMTRAK, for
instance.)  In that case is_cyhy_stakeholder needs to be set to true.

This commit adds code to perform the check against the CyHy database
and set is_cyhy_stakeholder accordingly.