Rework logic that adds data to collection
The previous code, although neater, had the issue that someone could
come along and grab the data between when the old rows were dropped
and the new rows were added.  In this new version we use update_one()
with upsert=True to update/create the rows.  Once that is complete we
delete any still-existing rows whose scan_date value is not the most
current.

Once we move to MongoDB 4 I can use a transaction to to atomically (1)
drop all the rows from the collection and (2) use insert_many() to
insert all the new data.  That will be much cleaner!